requirements.txt
flask
google-cloud-datastore
google-cloud-storage
gunicorn
wikipedia
